PM: Back off suspend if repeated attempts fail
If 10 attempts to suspend either fail or immediately resume within
1 second, back off suspending for 5 seconds multiplied by the
number of consecutive backoffs.  That is, pause suspends for 5
seconds after the first 10 failures, 10 seconds after the next
10 failures, and so on.

Avoid continually trying to suspend in situations in which a
driver is repeatedly rejecting suspend or a pending wakeup
interrupt is not handled, burning CPU in the continuous suspend
attempts.

Change-Id: I32289676d95a307ea3aa5e78f6c126ca979c0fec
Signed-off-by: Todd Poynor <toddpoynor@google.com>

Conflicts:

	kernel/power/wakelock.c